Added basic Plex search functionality. Before adding a movie request the backend will first check so that it is not already available on Plex.

+++ movie/api/views.py
@@ -15,6 +15,7 @@ from movie.models import Request
 from .serializers import MovieCreateSerializer, MovieListSerializer
 
 from .couchpotato import Couchpotato
+from frontpage_backend.plex import Plex
 
 
 class MovieCreateAPIView(APIView):
@@ -25,32 +26,40 @@ class MovieCreateAPIView(APIView):
 
         # Create a new Couchpotato object with the settings specified in settings.py
         couchpotato = Couchpotato(settings.COUCHPOTATO_HOST, settings.COUCHPOTATO_PORT, settings.COUCHPOTATO_API_KEY)
+        plex = Plex(settings.PLEX_HOST, settings.PLEX_PORT)
 
         try:
             # Get the session of the current user.
             session = SessionStore(session_key=request.data['sessionid'])
 
             try:
-                user = session['plexjocke_username']
-                # If adding a request for a movie succeeds.
-                if couchpotato.addmovie(request.data['imdb_id']):
+                token = session['plexjocke_username']
 
-                    # Remove the sessionid from the post data.
-                    data = request.data
-                    del data['sessionid']
+                if token:
+                    if not plex.search_for_movie(request.data['title'], request.data['release_date']):
+                        # If adding a request for a movie succeeds.
+                        if couchpotato.addmovie(request.data['imdb_id']):
 
-                    # Create a new MovieCreateSerializer using the data that was posted.
-                    serializer = MovieCreateSerializer(data=data)
+                            # Remove the sessionid from the post data.
+                            data = request.data
+                            del data['sessionid']
 
-                    # If the serializer is valid, save the request to the database and return HTTP status 201.
-                    if serializer.is_valid():
-                        serializer.save(cp_id = couchpotato.reply['movie']['_id'], user = session['plexjocke_username'], user_email = session['plexjocke_email'])
-                        return Response(serializer.data, status=status.HTTP_201_CREATED)
-                    # If the serializer failed return an error.
+                            # Create a new MovieCreateSerializer using the data that was posted.
+                            serializer = MovieCreateSerializer(data=data)
+
+                            # If the serializer is valid, save the request to the database and return HTTP status 201.
+                            if serializer.is_valid():
+                                serializer.save(cp_id = couchpotato.reply['movie']['_id'], user = session['plexjocke_username'], user_email = session['plexjocke_email'])
+                                return Response(serializer.data, status=status.HTTP_201_CREATED)
+                            # If the serializer failed return an error.
+                            else:
+                                return Response({'message': 'Movie has already been requested.', 'success': False}, status=status.HTTP_409_CONFLICT)
+                        else:
+                            return Response(couchpotato.reply, status=status.HTTP_503_SERVICE_UNAVAILABLE)
                     else:
-                        return Response({'message': 'Movie has already been requested.', 'success': False}, status=status.HTTP_409_CONFLICT)
+                        return Response({'message': 'Movie is already on Plex.', 'success': False}, status=status.HTTP_409_CONFLICT)    
                 else:
-                    return Response(couchpotato.reply, status=status.HTTP_503_SERVICE_UNAVAILABLE)
+                    return Response(status=status.HTTP_401_UNAUTHORIZED)
             except KeyError:
                 return Response(status=status.HTTP_401_UNAUTHORIZED)
 
@@ -67,19 +76,33 @@ class MovieDeleteAPIView(APIView):
         couchpotato = Couchpotato(settings.COUCHPOTATO_HOST, settings.COUCHPOTATO_PORT, settings.COUCHPOTATO_API_KEY)
 
         try:
-            # Fetch the movie to delete using the primary key specified.
-            movie = Request.objects.get(pk=pk)
+            # Get the session of the current user.
+            session = SessionStore(session_key=request.data['sessionid'])
 
-            # If Couchpotato successfully removed the movie.
-            if couchpotato.deletemovie(movie.cp_id):
-                # Delete the movie from the request database.
-                movie.delete()
+            try:
+                token = session['plexjocke_token']
 
-                return Response(couchpotato.reply, status=status.HTTP_204_NO_CONTENT)
-            else:
-                return Response(couchpotato.reply, status=status.HTTP_400_BAD_REQUEST)
-        except Request.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
+                if token:
+                    try:
+                        # Fetch the movie to delete using the primary key specified.
+                        movie = Request.objects.get(pk=pk)
+
+                        # If Couchpotato successfully removed the movie.
+                        if couchpotato.deletemovie(movie.cp_id):
+                            # Delete the movie from the request database.
+                            movie.delete()
+
+                            return Response(couchpotato.reply, status=status.HTTP_204_NO_CONTENT)
+                        else:
+                            return Response(couchpotato.reply, status=status.HTTP_400_BAD_REQUEST)
+                    except Request.DoesNotExist:
+                        return Response(status=status.HTTP_404_NOT_FOUND)
+                else:
+                    return Response(status=status.HTTP_401_UNAUTHORIZED)
+            except KeyError:
+                return Response(status=status.HTTP_401_UNAUTHORIZED)
+        except MultiValueDictKeyError:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
 
 
 class MovieDetailAPIView(RetrieveAPIView):
